Get testerUsername in taskattemptHistory

+++ src/main/java/com/itest/baseapplication/dto/TesterTaskAttemptDTO.java
@@ -15,7 +15,7 @@ import java.time.LocalDateTime;
 @Getter
 @Setter
 public class TesterTaskAttemptDTO {
-private Integer attemptId;
+private Long attemptId;
     
     private Integer taskId;
